+++ Assets/HUF/GenericDialog/Runtime/Configs/HGenericDialogConfig.cs
@@ -10,6 +10,7 @@ namespace HUF.GenericDialog.Runtime.Configs
         public const string CONFIG_ROOT_PATH = "HUF/GenericDialog/";
 
         public bool showSecondaryButton;
+        public bool showTertiaryButton;
         public HGenericDialogInstance prefab;
         public HGenericDialogShowSchema showSchema;
 
@@ -17,5 +18,14 @@ namespace HUF.GenericDialog.Runtime.Configs
         public string contentTranslation;
         public string primaryButtonTranslation;
         public string secondaryButtonTranslation;
+        public string tertiaryButtonTranslation;
+
+        public string HeaderTranslation => headerTranslation.Replace( "#COLOR#", $"#{ColorUtility.ToHtmlStringRGBA(linkColor)}" );
+        public string ContentTranslation => contentTranslation.Replace( "#COLOR#", $"#{ColorUtility.ToHtmlStringRGBA(linkColor)}" );
+        public string PrimaryButtonTranslation => primaryButtonTranslation.Replace( "#COLOR#", $"#{ColorUtility.ToHtmlStringRGBA(linkColor)}" );
+        public string SecondaryButtonTranslation => secondaryButtonTranslation.Replace( "#COLOR#", $"#{ColorUtility.ToHtmlStringRGBA(linkColor)}" );
+        public string TertiaryButtonTranslation => tertiaryButtonTranslation.Replace( "#COLOR#", $"#{ColorUtility.ToHtmlStringRGBA(linkColor)}" );
+
+        public Color linkColor;
     }
 }