+++ Assets/HUF/NotificationsFirebase/Runtime/API/HNotificationsFirebase.cs
@@ -1,8 +1,11 @@
 ﻿using System;
+using Firebase.Extensions;
 using Firebase.Messaging;
 using HUF.Notifications.Runtime.API;
 using HUF.NotificationsFirebase.Runtime.Implementation;
 using HUF.Utils.Runtime.Configs.API;
+using HUF.Utils.Runtime.Extensions;
+using HUF.Utils.Runtime.Logging;
 using JetBrains.Annotations;
 using UnityEngine;
 
@@ -10,6 +13,8 @@ namespace HUF.NotificationsFirebase.Runtime.API
 {
     public static class HNotificationsFirebase
     {
+        static readonly HLogPrefix logPrefix = new HLogPrefix( nameof(HNotificationsFirebase) );
+
         static FirebasePushNotificationsService firebaseNotifications;
 
         /// <summary>
@@ -68,6 +73,30 @@ namespace HUF.NotificationsFirebase.Runtime.API
                 firebaseNotifications.OnInitialized += HandleInitComplete;
             }
         }
+
+#if DEVELOPMENT_BUILD || UNITY_EDITOR
+        /// <summary>
+        /// Prints a FCM token in the console and returns it in the callback.
+        /// Can be used for sending test push messages to a single device.
+        /// </summary>
+        /// <param name="callback">An operation callback.</param>
+        [PublicAPI]
+        public static void FetchFCMToken( Action<string> callback )
+        {
+            FirebaseMessaging.GetTokenAsync().ContinueWithOnMainThread( task =>
+            {
+                if ( !task.IsCompleted || string.IsNullOrEmpty( task.Result ) )
+                {
+                    HLog.LogWarning( logPrefix, "Retrieving FCM token failed" );
+                    callback.Dispatch( null );
+                    return;
+                }
+
+                HLog.LogAlways( logPrefix, $"FCM Token:\n{task.Result}" );
+                callback.Dispatch( task.Result );
+            } );
+        }
+#endif
 
         [RuntimeInitializeOnLoadMethod( RuntimeInitializeLoadType.BeforeSceneLoad )]
         static void AutoInit()